feat: sync reset and set idb and stores
