+++ src/frontend/src/lib/services/idb-store.services.ts
@@ -1,0 +1,58 @@
+import type { CanisterStore } from '$lib/stores/canister.store';
+import type { CanisterIdText } from '$lib/types/canister';
+import { del, entries as idbEntries, set, type UseStore } from 'idb-keyval';
+
+export const syncIdbStore = async <T>({
+	customStore,
+	store
+}: {
+	customStore: UseStore;
+	store: CanisterStore<T>;
+}) => {
+	const entries = await idbEntries<CanisterIdText, T>(customStore);
+
+	store.setAll(
+		entries.length === 0
+			? undefined
+			: entries.reduce(
+					(acc, [key, value]) => ({
+						...acc,
+						[key]: value
+					}),
+					{}
+				)
+	);
+};
+
+export const setIdbStore = async <T>({
+	customStore,
+	store,
+	canisterId,
+	data
+}: {
+	customStore: UseStore;
+	store: CanisterStore<T>;
+	canisterId: CanisterIdText;
+	data: T;
+}) => {
+	store.set({
+		canisterId,
+		data
+	});
+
+	await set(canisterId, data, customStore);
+};
+
+export const resetIdbStore = async <T>({
+	customStore,
+	store,
+	canisterId
+}: {
+	customStore: UseStore;
+	store: CanisterStore<T>;
+	canisterId: CanisterIdText;
+}) => {
+	store.reset(canisterId);
+
+	await del(canisterId, customStore);
+};
