+++ src/frontend/src/routes/+layout.svelte
@@ -7,13 +7,14 @@
 	import Spinner from '$lib/components/ui/Spinner.svelte';
 	import { displayAndCleanLogoutMsg, signOut } from '$lib/services/auth.services';
 	import { initMissionControl } from '$lib/services/console.services';
+	import { syncSubnets } from '$lib/services/ic.services';
+	import { syncSnapshots } from '$lib/services/snapshots.services';
 	import { initAuthWorker } from '$lib/services/worker.auth.services';
 	import { type AuthStoreData, authStore } from '$lib/stores/auth.store';
 	import { i18n } from '$lib/stores/i18n.store';
 	import { missionControlStore } from '$lib/stores/mission-control.store';
 	import { toasts } from '$lib/stores/toasts.store';
 	import '$lib/styles/global.scss';
-	import { syncCanisterStores } from '$lib/services/canister.services';
 
 	interface Props {
 		children: Snippet;
@@ -21,7 +22,8 @@
 
 	let { children }: Props = $props();
 
-	const init = async () => await Promise.all([i18n.init(), syncAuthStore(), syncCanisterStores()]);
+	const init = async () =>
+		await Promise.all([i18n.init(), syncAuthStore(), syncSubnets(), syncSnapshots()]);
 
 	const syncAuthStore = async () => {
 		if (!browser) {
